pyproject.toml
[project]
name = "content-farm"
version = "0.1.0"
description = "Automated content pipeline from Reddit to short-form video"
requires-python = ">=3.11"
dependencies = [
    "langgraph>=0.2.0",
    "langchain-groq>=0.2.0",
    "httpx>=0.27.0",
    "rich>=13.0.0",
    "edge-tts>=6.1.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "browser-use>=0.11.9",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "UP"]
